test: add coverage for gitignore warnings, extract to separate file
Extract gitignore-related runner tests into runner-gitignore.test.ts to
stay under the 500-line max-lines lint rule. Add 3 new tests covering:
- warns when respectGitignore is false without files scope
- warns when file patterns match only gitignored files
- does not warn when file patterns match tracked files
